feat(services): Adds pagination, adds desired interface attributes & removes caching
The removal of caching is temporary whilst I try to figure out how to run HTTPS locally to properly test it (as it's currently broken). However, I also would like to implement proper error handling and so on. Otherwise, this is a great start. Additionally, the use of Set - which I learnt about recently - makes sure that each entry is unique, avoiding a duplicate entry appearing on another page.